{curs}{26.11.2023_11:59}{Работа с возможностью переключения со страницы персональной на уроки курса и на окончания курса}

+++ src/main/java/com/example/servercurs/entities/Group.java
@@ -53,6 +53,6 @@ public class Group {
     private TimeTable timetable;
 
     @Transient
-    private int progress;
+    private double progress;
 
 }
